requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
# cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
# cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
inflection==0.5.1
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.7
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
zope.interface==6.4.post2
